2019 Day 9 Part 1
Relative bases, a new parameter mode, 64-bit ints...oh my

+++ 2019/channeltypes/types.go
@@ -1,8 +1,8 @@
 package channeltypes
 
 // List makes a channel that just outputs the requested list
-func List(items ...int) <-chan int {
-	out := make(chan int, len(items))
+func List(items ...int64) <-chan int64 {
+	out := make(chan int64, len(items))
 	for _, v := range items {
 		out <- v
 	}
@@ -11,8 +11,8 @@ func List(items ...int) <-chan int {
 }
 
 // Chain makes a channel that pulls from each input channel in sequence
-func Chain(inputs ...<-chan int) <-chan int {
-	out := make(chan int, 1)
+func Chain(inputs ...<-chan int64) <-chan int64 {
+	out := make(chan int64, 1)
 	impl := func() {
 		for _, c := range inputs {
 			for v := range c {
